modularize code - move voice commands into its own service file
